refactor(Graph): Create graph connected component manager
